+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.2]
+### Added
+- Ideoplots - function to draw ideoplots from user provided data
+- JSCode class for javascript code snippets
+
 ## [v0.2.1]
 ### Fixed
 - conda requirements